bug involving gzread and its stupid unisgned int argument, as opposed to size_t
git-svn-id: svn+ssh://combine-101/home/svn_bio/shrimp/trunk@689 450597fd-3c47-0410-84d8-f85f2794411d

+++ common/util.h
@@ -109,7 +109,7 @@ char	       *fast_gzgets(gzFile, char*, int);
 char	       *fast_gzgets_safe(fasta_t f);
 char	       *comma_integer(uint64_t);
 void xgzwrite(gzFile fp, voidp buf, unsigned len);
-void xgzread(gzFile fp, voidp buf, unsigned len);
+void xgzread(gzFile fp, voidp buf, size_t len);
 
 
 /* for optarg (and to shut up icc) */
